requirements.txt
apache-airflow==2.4.0
apache-airflow-providers-amazon==6.0.0
apache-airflow-providers-common-sql==1.2.0
apache-airflow-providers-ftp==3.1.0
apache-airflow-providers-http==4.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-postgres==5.2.0
apache-airflow-providers-sqlite==3.2.1
pandas==1.3.5
python-decouple==3.6
Jinja2==3.1.2
Unidecode==1.3.6